# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FpolyZone",
        "version": "0.1"
    },
    "paths": {
...
        "/api/admin/user/detail/{user}": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Xem chi tiết tài khoản người dùng",
                "description": "Xem chi tiết thông tin tài khoản người dùng.",
                "operationId": "438eb505190666bbc07a7ad55613aac8",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID của tài khoản người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin chi tiết của tài khoản người dùng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID tài khoản người dùng",
                                            "type": "integer",
                                            "format": "int64"
                                        },
                                        "username": {
                                            "description": "Tên đăng nhập",
                                            "type": "string"
                                        },
                                        "first_name": {
                                            "description": "Họ",
                                            "type": "string"
                                        },
                                        "last_name": {
                                            "description": "Tên",
                                            "type": "string"
                                        },
                                        "group_id": {
                                            "description": "ID nhóm",
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "email": {
                                            "description": "Địa chỉ email",
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "birthday": {
                                            "description": "Ngày sinh",
                                            "type": "string"
                                        },
                                        "avatar": {
                                            "description": "Link đến ảnh đại diện",
                                            "type": "string"
                                        },
                                        "phone": {
                                            "description": "Số điện thoại",
                                            "type": "string"
                                        },
                                        "address": {
                                            "description": "Địa chỉ",
                                            "type": "string"
                                        },
                                        "biography": {
                                            "description": "Tiểu sử",
                                            "type": "string"
                                        },
                                        "gender": {
                                            "description": "Giới tính",
                                            "type": "string"
                                        },
                                        "status": {
                                            "description": "Trạng thái",
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "major_id": {
                                            "description": "ID ngành học",
                                            "type": "integer",
                                            "format": "int32"
                                        },
                                        "permissions": {
                                            "description": "Danh sách quyền",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "verification_code": {
                                            "description": "Mã xác minh",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Ngày tạo",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Ngày cập nhật",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy tài khoản"
                    }
                }
            }
        },
        "/api/admin/user/suspend/{user}": {
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "Tạm khóa tài khoản người dùng",
                "description": "Tạm khóa tài khoản người dùng dựa trên ID.",
                "operationId": "858b4c93688c2026457c1ff85d4d1e8e",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID của tài khoản người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tạm khóa tài khoản thành công"
                    },
                    "400": {
                        "description": "Lỗi trong quá trình xử lý"
                    },
                    "404": {
                        "description": "Không tìm thấy tài khoản"
                    }
                }
            }
        },
        "/api/admin/user/active/{user}": {
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "Mở khóa tài khoản người dùng",
                "description": "Mở khóa tài khoản người dùng dựa trên ID.",
                "operationId": "a28a63c9aaa7014b0b82c2ec63442b8e",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID của tài khoản người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mở khóa tài khoản thành công"
                    },
                    "400": {
                        "description": "Lỗi trong quá trình xử lý"
                    },
                    "404": {
                        "description": "Không tìm thấy tài khoản"
                    }
                }
            }
        },
        "/api/admin/user/delete/{user}": {
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "Xóa tài khoản người dùng",
                "description": "Xóa tài khoản người dùng dựa trên ID.",
                "operationId": "f8ece4233a9eae56293e927d468291b0",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID của tài khoản người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa tài khoản thành công"
                    },
                    "400": {
                        "description": "Lỗi trong quá trình xử lý"
                    },
                    "404": {
                        "description": "Không tìm thấy tài khoản"
                    }
                }
            }
        },
...
    }
}